Added a Blender Addon script. The addon has a preferences dropdown to file-select the BlenderProvider path. Updated the README with setup instructions
