+++ README.md
@@ -2,4 +2,14 @@
 
 This is the CodeToCAD Blender Provider repo, that allows CodeToCAD to talk to Blender.
 
-Please refer to the [CodeToCAD repo](https://github.com/CodeToCad/CodeToCAD) for more information about this tool.
+Please refer to the [CodeToCAD repo](https://github.com/CodeToCad/CodeToCAD) for more information about this CodeToCAD.
+
+## Setup
+
+The easiest way to use this BlenderProvider is to enable the addon in Blender:
+
+1. In Blender, go to Edit -> Preferences -> Add-ons. 
+2. Click Install. Select the [CodeToCADBlenderAddon.py](./CodeToCADBlenderAddon.py) file.
+3. Enable the addon in blender by clicking on the checkmark.
+4. Expand the arrow next to the CodeToCAD addon, this will open up the preferences.
+5. Select the BlenderProvider folder, which is the root of this repo.
